Fix Dash feature in levels 3 and 4
This commit resolves an issue with the Dash feature not working in levels 3 and 4. The problem was due to the use of a different player prefab (Prefabs/Player.prefab) in these levels, which did not have the necessary Trail Renderer for the Dash feature.

The solution was to replace the player prefab in levels 3 and 4 with the original prefab (Prefabs/Characters/Player.prefab) used in levels 1 and 2. This prefab has the Dash feature properly implemented.

If the Prefabs/Player.prefab is not used elsewhere, it is recommended to remove it after verification to avoid similar issues in the future.